+++ pegasus-game/core/src/com/tlear/pegasus/ShipWeapon.java
@@ -4,21 +4,21 @@ import com.badlogic.gdx.graphics.g2d.TextureRegion;
 
 public interface ShipWeapon {
 	
-	public void setX(double x);
-	public void setY(double y);
+	public void setX(float x);
+	public void setY(float y);
 	public void setAngle(float a);
 	
-	public void addX(double x);
-	public void addY(double y);
+	public void addX(float x);
+	public void addY(float y);
 	public void addAngle(float a);
 	
-	public double getX();
-	public double getY();
+	public float getX();
+	public float getY();
 	public float getAngle();
 	public int getWidth();
 	public int getHeight();
-	public int getTexWidth();
-	public int getTexHeight();
+	public float getTexWidth();
+	public float getTexHeight();
 	
 	public TextureRegion getTextureRegion();
